Add basic and auto themes, clean up css files.
Change the `kittygram-light` and `kittygram-dark` css files to only contain the changed variables.
Move the rest of the css to a `base.css` file.

Add an `auto` theme that changes color scheme based on a user's preference.

Move the `themes` table into it's own file.

Add a `basic` theme (should have compatibility with all browsers).

Change the settings icon to one that looks nicer.